+++ src/components/productDetail-page/productInfo.tsx/ProductInfo.tsx
