+++ src/components/productDetail-page/productInfo.tsx/totalPrice/TotalPrice.tsx
